+++ example/android/app/src/androidTest/java/com/instabug/react/example/DetoxTest.java
@@ -1,4 +1,4 @@
-package com.instabugsample;
+package com.instabug.react.example;
 
 import com.wix.detox.Detox;
 import com.wix.detox.config.DetoxConfig;
@@ -23,7 +23,7 @@ public class DetoxTest {
         DetoxConfig detoxConfig = new DetoxConfig();
         detoxConfig.idlePolicyConfig.masterTimeoutSec = 90;
         detoxConfig.idlePolicyConfig.idleResourceTimeoutSec = 60;
-        detoxConfig.rnContextLoadTimeoutSec = (com.instabugsample.BuildConfig.DEBUG ? 180 : 60);
+        detoxConfig.rnContextLoadTimeoutSec = (com.instabug.react.example.BuildConfig.DEBUG ? 180 : 60);
 
         Detox.runTests(mActivityRule, detoxConfig);
     }
